Restructuring:  Making `plot.py`
Restructuring to clean up simulation.py and better organize our package.

- Moved all plotting functions from `simulation.py` to a new file `plot.py`.
- From `simulation.py`, moved the `apply_rot_pulse` function to `operators.py`, where it is more appropriate.

In the future:

- We should eventually replace all use of the plotting functions with `pulsee.plot` instead of `pulsee.simulation` for better organization.
For now, I have imported all the plotting functions into `simulation.py` so that all demos will run correctly.
But eventually, we should go through all demos & example files and update the use of plotting functions.

- Replace the use of matplotlib.pylab (strongly discouraged by matplotlib documentation) with matplotlib.pyplot
